+++ releasenotes.md
@@ -1,5 +1,6 @@
 Whats New in v8.5
 --------------------------
+- Fix formulas json export
 - Ingredient general tab re-design
 - Pubchem tab re-design
 - Adding more details when a note is added during formula making